Refactored permission granting into trait
